Add github action to sync with public repo (#15)
This PR adds a github action workflow that automatically syncs the
internal with the public repository.

I've already rebased the internal repo to include all the current
commits from the public repo.

This makes a push with --force, so all changes on the public main that
aren't done in the main are overwritten!

Base64 decoded private key is already added to the github actions
secrets.